+++ drift_dev/lib/api/migrations.dart
@@ -73,6 +73,21 @@ abstract class SchemaVerifier {
   /// expected exist.
   Future<void> migrateAndValidate(GeneratedDatabase db, int expectedVersion,
       {bool validateDropped = false});
+
+  /// Utility function used by generated tests to verify that migrations
+  /// modify the database schema as expected.
+  ///
+  /// Foreign key constraints are disabled for this operation.
+  Future<void> testWithDataIntegrity<OldDatabase extends GeneratedDatabase,
+      NewDatabase extends GeneratedDatabase>({
+    required OldDatabase Function(QueryExecutor) createOld,
+    required NewDatabase Function(QueryExecutor) createNew,
+    required GeneratedDatabase Function(QueryExecutor) openTestedDatabase,
+    required void Function(Batch, OldDatabase) createItems,
+    required Future Function(NewDatabase) validateItems,
+    required int oldVersion,
+    required int newVersion,
+  });
 }
 
 /// Utilities verifying that the current schema of the database matches what