+++ drift_dev/lib/src/services/schema/verifier_impl.dart
@@ -100,6 +100,31 @@ class VerifierImplementation implements SchemaVerifier {
   Future<DatabaseConnection> startAt(int version) {
     return schemaAt(version).then((schema) => schema.newConnection());
   }
+
+  @override
+  Future<void> testWithDataIntegrity<OldDatabase extends GeneratedDatabase,
+          NewDatabase extends GeneratedDatabase>(
+      {required OldDatabase Function(QueryExecutor p1) createOld,
+      required NewDatabase Function(QueryExecutor p1) createNew,
+      required GeneratedDatabase Function(QueryExecutor p1) openTestedDatabase,
+      required void Function(Batch p1, OldDatabase p2) createItems,
+      required Future Function(NewDatabase p1) validateItems,
+      required int oldVersion,
+      required int newVersion}) async {
+    final schema = await schemaAt(oldVersion);
+
+    final oldDb = createOld(schema.newConnection());
+    await oldDb.batch((batch) => createItems(batch, oldDb));
+    await oldDb.close();
+
+    final db = openTestedDatabase(schema.newConnection());
+    await migrateAndValidate(db, newVersion);
+    await db.close();
+
+    final newDb = createNew(schema.newConnection());
+    await validateItems(newDb);
+    await newDb.close();
+  }
 }
 
 Input? _parseInputFromSchemaRow(